Edited cleanup

+++ cleanup.sh
@@ -1,1 +1,9 @@
-/usr/bin/hadoop fs -rm -r /user/wordcount
+testPlatform="Amazon"
+tester="tgreenside"
+
+if ["$testPlatform" == "Amazon"]
+then
+    /usr/bin/hadoop fs -rm -r /user/wordcount
+else
+    hadoop fs -rmr /user/"$tester"
+fi
